requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
langchain==0.3.27
langchain-google-genai==2.1.8
langchain-community==0.3.29
faiss-cpu==1.12.0
pypdf==5.9.0
docx2txt==0.9
python-dotenv==1.1.1
tiktoken==0.11.0
